Week_6/demo-1-code-review-autogen/requirements.txt
autogen-agentchat>=0.4.0
autogen-ext[openai]>=0.4.0
click>=8.1.0
rich>=13.0.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
